requirements.txt
boto3==1.20.5
python-dotenv==0.19.2
